Move diff and countMatchBlock to diff_helpers_test.go
